privatize TestAssetBundle, move multi_screen_golden_file to its own file
