+++ packages/golden_toolkit/lib/src/multi_screen_golden.dart
@@ -1,5 +1,11 @@
+/// ***************************************************
+/// Copyright 2019-2020 eBay Inc.
+///
+/// Use of this source code is governed by a BSD-style
+/// license that can be found in the LICENSE file or at
+/// https://opensource.org/licenses/BSD-3-Clause
+/// ***************************************************
 import 'dart:ui';
-
 import 'package:flutter_test/flutter_test.dart';
 import 'testing_tools.dart';
