+++ packages/golden_toolkit/lib/src/test_asset_bundle.dart
@@ -1,0 +1,32 @@
+import 'dart:convert';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/services.dart';
+
+/// ***************************************************
+/// Copyright 2019-2020 eBay Inc.
+///
+/// Use of this source code is governed by a BSD-style
+/// license that can be found in the LICENSE file or at
+/// https://opensource.org/licenses/BSD-3-Clause
+/// ***************************************************
+///
+
+/// TestAssetBundle is required in order to avoid issues with large assets
+///
+/// ref: https://medium.com/@sardox/flutter-test-and-randomly-missing-assets-in-goldens-ea959cdd336a
+///
+class TestAssetBundle extends CachingAssetBundle {
+  @override
+  Future<String> loadString(String key, {bool cache = true}) async {
+    //overriding this method to avoid limit of 10KB per asset
+    final ByteData data = await load(key);
+    if (data == null) {
+      throw FlutterError('Unable to load asset, data is null: $key');
+    }
+    return utf8.decode(data.buffer.asUint8List());
+  }
+
+  @override
+  Future<ByteData> load(String key) async => rootBundle.load(key);
+}
